nnue: hold accumulators in int16 instead of int32 (#146)
The feature transformer's weights are int16 and its accumulator was int32,
so every incremental update moved twice the bytes it needed to and filled
half as many AVX2 lanes per vector. That is the hot path: lowering QA to 127
so the dense layers could use maddubs bought only 7%, which said the dense
layers were never the bottleneck -- the L1=256 accumulator was.

Safety is a property of the weights, so it is checked at load rather than
assumed. HalfKP activates one feature per non-king piece, so thirty is the
most that can ever sum into one slot; the loader rejects any network whose
feature weights could carry it past int16. The current net's worst case is
+/-1288 against a range of +/-32767, so the bound is not close.

The evaluation is unchanged, and that is checked rather than argued: 199/199
tests pass, including the reference-vs-AVX2 equality over 20,000 random
accumulators and the incremental-vs-full-recompute check, and 200 book
positions searched to depth 9 return identical bestmoves.

Bench depth 11, 5-rep medians, nps:

              single-threaded   24-way concurrent (per instance)
  HCE            1,906,803        1,239,751
  int32          1,116,725          514,620
  int16          1,800,920          817,981

The penalty against HCE falls from 41.4% to 5.6% single-threaded. The second
column is the reason this matters more than it first looks: measured alone
the int32 penalty is 41.4%, but under the 24-way concurrency the matches
actually run at, it is 58.5%. Every NNUE speed figure quoted so far was taken
on an idle box and understated the cost of a real match.

+++ include/havoc/eval/nnue_evaluator.hpp
@@ -52,7 +52,7 @@ class NNUEEvaluator : public IEvaluator {
     [[nodiscard]] int evaluate(const position& pos, int lazy_margin = -1) override {
         (void)lazy_margin;
         const int us = static_cast<int>(pos.to_move());
-        const int32_t* base = frame(top_);
+        const int16_t* base = frame(top_);
         return net_->forward(base + static_cast<size_t>(us) * static_cast<size_t>(l1_),
                              base + static_cast<size_t>(us ^ 1) * static_cast<size_t>(l1_));
     }
@@ -70,7 +70,7 @@ class NNUEEvaluator : public IEvaluator {
 
     void refresh(const position& pos) override {
         top_ = 0;
-        int32_t* base = frame(0);
+        int16_t* base = frame(0);
         recompute(base, pos, white);
         recompute(base + static_cast<size_t>(l1_), pos, black);
     }
@@ -93,15 +93,15 @@ class NNUEEvaluator : public IEvaluator {
 
     [[nodiscard]] size_t frame_stride() const { return static_cast<size_t>(2 * l1_); }
 
-    [[nodiscard]] int32_t* frame(int i) {
+    [[nodiscard]] int16_t* frame(int i) {
         return frames_.data() + static_cast<size_t>(i) * frame_stride();
     }
-    [[nodiscard]] const int32_t* frame(int i) const {
+    [[nodiscard]] const int16_t* frame(int i) const {
         return frames_.data() + static_cast<size_t>(i) * frame_stride();
     }
 
     /// Rebuild one perspective from the board. `dst` is `l1_` wide.
-    void recompute(int32_t* dst, const position& pos, Color perspective) const {
+    void recompute(int16_t* dst, const position& pos, Color perspective) const {
         const int16_t* bias = net_->ft_bias();
         for (int i = 0; i < l1_; ++i)
             dst[i] = bias[i];
@@ -116,7 +116,7 @@ class NNUEEvaluator : public IEvaluator {
     int l1_ = 0;
     int top_ = 0;
     int capacity_ = 0;
-    std::vector<int32_t> frames_;
+    std::vector<int16_t> frames_;
 };
 
 inline void NNUEEvaluator::push(const position& pos, const FeatureDelta& d) {
@@ -124,9 +124,9 @@ inline void NNUEEvaluator::push(const position& pos, const FeatureDelta& d) {
         capacity_ *= 2;
         frames_.resize(static_cast<size_t>(capacity_) * frame_stride(), 0);
     }
-    const int32_t* prev = frame(top_);
+    const int16_t* prev = frame(top_);
     ++top_;
-    int32_t* cur = frame(top_);
+    int16_t* cur = frame(top_);
 
     // A king move rebuilds only its own perspective. Detect it before touching
     // anything, because the events for the two perspectives are the same list
@@ -138,13 +138,13 @@ inline void NNUEEvaluator::push(const position& pos, const FeatureDelta& d) {
 
     for (int ci = 0; ci < 2; ++ci) {
         const Color perspective = static_cast<Color>(ci);
-        int32_t* dst = cur + static_cast<size_t>(ci) * static_cast<size_t>(l1_);
+        int16_t* dst = cur + static_cast<size_t>(ci) * static_cast<size_t>(l1_);
         if (king_moved[ci]) {
             recompute(dst, pos, perspective);
             continue;
         }
         std::memcpy(dst, prev + static_cast<size_t>(ci) * static_cast<size_t>(l1_),
-                    static_cast<size_t>(l1_) * sizeof(int32_t));
+                    static_cast<size_t>(l1_) * sizeof(int16_t));
 
         // This perspective's king did not move, so its bucket is the same
         // before and after -- reading it from the post-move board is safe.
@@ -164,10 +164,10 @@ inline void NNUEEvaluator::push(const position& pos, const FeatureDelta& d) {
 }
 
 inline bool NNUEEvaluator::matches_full_recompute(const position& pos) const {
-    std::vector<int32_t> want(static_cast<size_t>(l1_));
+    std::vector<int16_t> want(static_cast<size_t>(l1_));
     for (int ci = 0; ci < 2; ++ci) {
         recompute(want.data(), pos, static_cast<Color>(ci));
-        const int32_t* have = frame(top_) + static_cast<size_t>(ci) * static_cast<size_t>(l1_);
+        const int16_t* have = frame(top_) + static_cast<size_t>(ci) * static_cast<size_t>(l1_);
         for (int i = 0; i < l1_; ++i)
             if (have[i] != want[i])
                 return false;
